Added folders src, img, docs, .vscode. Updated Cpp.hpp. Added LICENSE and documentation
