+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "files.associations": {
+        "concepts": "cpp",
+        "type_traits": "cpp",
+        "xtr1common": "cpp",
+        "string" : "cpp"
+    }
+}
