--- Sposzczone-jezyki-programowania/Spolszczone-jezyki-programowania/VCPP/Funkcje.cpp
@@ -1,10 +1,0 @@
-#include <string>
-using namespace std;
-namespace Funkcje
-{
-	template <typename T>
-	static T Zwroc(T arg)
-	{
-		return arg;
-	}
-};
