+++ docs/LICENSE.md
@@ -1,0 +1,66 @@
+# Publiczna Licencja Spolszczonych Języków Programowania (PL SJP)
+## Na podstawie MsPL (Microsoft Public License)
+## Autor: Antoni Kiedos <antek_kiedos@outlook.com> (C) 2020
+
+Niniejsza licencja reguluje korzystanie z towarzyszącego oprogramowania. W przypadku korzystania z
+oprogramowania, użytkownik akceptuje niniejszą licencję. Jeśli nie akceptujesz licencji, nie
+korzystania z oprogramowania.
+
+### 1. Definicje
+- Terminy "kopiować", "kopia", "utwory pochodne" i "dystrybucja" mają tu takie samo znaczenie jak na mocy polskiego prawa autorskiego. 
+- "Wkład" jest oryginalnym oprogramowaniem lub jakimikolwiek dodatkami lub zmianami w oprogramowaniu.
+- "Współautorem" jest każda osoba, która dystrybuuje swój wkład w ramach Licencji.
+- "Patenty licencjonowane" są oświadczeniami patentowymi bezpośrednio od jego wkładu.
+
+### 2. Przyznanie praw
+
+#### (A)
+        
+    Udzielenie praw autorskich - z zastrzeżeniem warunków niniejszej licencji, w tym
+    warunków licencji i ograniczeń w sekcji 3, każdy podmiot przekazuje
+    niewyłączną, ogólnoświatową, nieodpłatną licencję praw autorskich, aby
+    odtworzyć swój wkład, przygotować prace pochodne swojego wkładu,
+    i rozpowszechniać swój wkład lub wszelkie utwory pochodne, które tworzy.
+
+#### (B)
+
+    Udzielenie patentu - z zastrzeżeniem warunków niniejszej licencji, w tym
+    warunków licencji i ograniczeń w sekcji 3, każdy podmiot przekazuje
+    niewyłączną, ogólnoświatową, nieodpłatną licencję na licencji
+    patentów, dokonywać, wykorzystywać, sprzedawać, oferować na sprzedaż, importować i/lub
+    w inny sposób zbywać swój wkład w oprogramowanie lub prace pochodne
+    wkładu w oprogramowanie.
+
+### 3. Warunki i ograniczenia
+    
+#### (A)
+
+    Brak licencji na znak towarowy - niniejsza licencja nie daje użytkownikowi praw do
+    nazwy, logo lub znaków towarowych współautorów.
+    
+#### (B)
+    
+    W przypadku modyfikacji kodu źródłowego do celów dystrybucji, zakazanym jest kopiowanie tegoż kodu oraz zmienianie nazw występujących w kodzie,
+    a także podmieniania jakichkolwiek innych słów, ani w komentarzach, ani w strumieniach komunikacji z użytkownikiem i w tak zmienionej formie dystrybuowanie.
+    Natomiast dozwolonym jest rozpowszechnianie kodu źródłowego zainspirowanego zawartością kodu źródłowego objętego tą Licencją pod warunkiem dopisania w dokumentacji lub komentarzach lub w strumieniach wyjścia informacji o postaci następującej:
+`"Na podstawie programu <Tutaj nazwa> stworzonego przez <Imię i nazwisko autora/autorów> <Rok wydania wersji oprogramowania>"`.
+    
+#### (C)
+    
+    W przypadku rozpowszechniania jakiejkolwiek części oprogramowania należy zachować wszelkie informacje o prawach autorskich,
+    patentach, znakach towarowych i atrybucji, które są obecne w oprogramowaniu.
+
+#### (D)
+    
+    W przypadku rozpowszechniania jakiejkolwiek części oprogramowania w formie kodu źródłowego,
+    można to zrobić wyłącznie na podstawie niniejszej licencji, dołączając pełną kopię
+    tej licencji wraz z dystrybucją. W przypadku dystrybucji jakiejkolwiek części
+    w formie skompilowanej lub w formie kodu obiektowego, można to zrobić tylko w ramach
+    licencji zgodnej z niniejszą licencją.
+
+#### (E)
+    
+    Oprogramowanie jest licencjonowane "tak jak jest". Ponosisz ryzyko jego użycia.
+    Autorzy nie udzielają wyraźnych gwarancji. Użytkownik może mieć dodatkowe prawa konsumenckie zgodnie z lokalnymi przepisami, których ta licencja nie może zmienić.
+    W zakresie dozwolonym przez lokalne przepisy, autorzy wykluczają dorozumiane gwarancje przydatności handlowej,
+    przydatności do określonego celu i nienaruszania praw.
