+++ docs/PPL PL.md
@@ -1,0 +1,66 @@
+# Polished Programming Languages Public License (PPL PL)
+## Based on MsPL (Microsoft Public License)
+## Autor: Antoni Kiedos <antek_kiedos@outlook.com> (C) 2020
+
+
+This License governs use of the accompanying software. If you use the
+software, you accept this License. If you do not accept the License, do not
+use the software.
+
+### 1.  Definitions
+- The terms "reproduce," "reproduction," "derivative works," and "distribution" have the same meaning here as under Polish copyright law.
+- A "contribution" is the original software, or any additions or changes to the software.
+- A "contributor" is any person that distributes its contribution under this License.
+- "Licensed patents" are a contributor's patent claims that read directly on its contribution.
+
+### 2.  Grant of Rights
+
+#### (A)
+    
+    Copyright Grant- Subject to the terms of this License, including the
+    License conditions and limitations in section 3, each contributor grants
+    you a non-exclusive, worldwide, royalty-free copyright License to
+    reproduce its contribution, prepare derivative works of its contribution,
+    and distribute its contribution or any derivative works that you create.
+
+#### (B)
+    
+    Patent Grant - Subject to the terms of this License, including the
+    License conditions and limitations in section 3, each contributor grants
+    you a non-exclusive, worldwide, royalty-free License under its license
+    patents to make, have made, use, sell, offer for sale, import, and/or
+    otherwise dispose of its contribution in the software or derivative works
+    of the contribution in the software.
+
+### 3.  Conditions and Limitations
+    
+#### (A)
+
+    No Trademark License - This License does not grant you rights to use
+    any contributors' name, logo, or trademarks.
+
+#### (B)
+    
+    If you modify the source code for distribution purposes, it is forbidden to copy the source code and change the names in the code, or to replace any other words, either in comments or in streams of communication with you and in such altered form as distributed. On the other hand, it is permitted to distribute source code inspired by the content of the source code under this License, provided that the following information is included in the documentation or comments or output streams:
+`"Based on the <Here program's name> program created by <Author's/authors' name(s)> <Software release date>"`.
+<!--
+
+-->
+#### (C)
+    
+    If you distribute any portion of the software, you must retain all copyright, patent, trademark, and attribution notices that are present in the software.
+
+#### (D)
+    
+    If you distribute any portion of the software in source code form,
+    you may do so only under this License by including a complete copy of
+    this License with your distribution. If you distribute any portion of the
+    software in compiled or object code form, you may only do so under a
+    License that complies with this License.
+
+#### (E)
+
+    The software is license "as-is." You bear the risk of using it. The contributors give no express warranties,
+    guarantees, or conditions. You may have additional consumer rights under your local laws which this License cannot change.
+    To the extent permitted under your local laws, the contributors exclude the implied warranties of merchantability,
+    fitness for a particular purpose and non-infringement.
