+++ docs/README.md
@@ -1,0 +1,36 @@
+---
+Author: Antoni Kiedos
+Project: PLC++
+---
+# Spis Treści
+* [1. Streszczenie](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Streszczenie)
+* [1.1. Cpp.hpp](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Cpp.hpp)
+  * [1.1.1. Typy Podstawowe](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Typy%2BPodstawowe)
+* [1.2. Funkcje.hpp](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Funkcje.cpp)
+  * [1.2.1. Funkcje::Zwroc()](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Funkce::Zwroc())
+* [1.3. Mat.cpp](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Mat.cpp)
+  * [1.3.1. Średnie](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/docs/README.md#Średnie)
+# Streszczenie
+## Patrz również: [Streszczenie streszczenia](https://github.com/Programista-Cpp/Sposzczone-jezyki-programowania/tree/master/README.md)
+## Cpp.hpp <!-- omit in toc -->
+> [!NOTE]
+> Jeśli masz dostęp do eksperymentalnej wersji standardu C++2b, w której jest:
+> * std::ranges::accumulate();
+> * lub std::colony;
+- [x] task
+Musisz zdefiniować
+
+> ```cpp
+> #define __CPP_2b
+> ```
+### Typy Podstawowe
+
+
+## Funkcje.cpp
+
+### Funkcje::Zwroc()
+
+
+## Mat.cpp
+
+### Średnie
