+++ src/PLCPP/pl_cpp_ext/Funkcje.cpp
@@ -1,0 +1,10 @@
+#include <string>
+using namespace std;
+namespace Funkcje
+{
+	template <typename T>
+	static T Zwroc(T arg)
+	{
+		return arg;
+	}
+};
